Cast double to float
Time is a double when gotten from glfw. To make the compile be happy I cast it to a float.

+++ Engine/src/Moon/Core/Application/Application.cpp
@@ -30,7 +30,7 @@ namespace Moon {
 		while (m_Running)
 		{
 			// Get delta time
-			float time = glfwGetTime();	// Platfrom::GetTime();
+			float time = (float)glfwGetTime();	// Platfrom::GetTime();
 			m_Timestep = time - m_LastFrameTime;
 			m_LastFrameTime = time;
